% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_marathon}
\alias{london_marathon}
\title{London marathon data set.}
\format{
A tibble with 42 rows and 8 variables:
\describe{
\item{Date}{Date of race}
\item{Year}{Year of race}
\item{Applicants}{Number of people who applied}
\item{Accepted}{Number of people accepted}
\item{Starters}{Number of people who started}
\item{Finishers}{Number of people who finished}
\item{Raised}{amount of money raised}
\item{Official charity}{Name of official charity}
}
}
\source{
\url{https://en.wikipedia.org/wiki/London_Marathon}
}
\usage{
london_marathon
}
\description{
A data set containing data on the participants and charities of London Marathon.
}
\keyword{datasets}
