% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winners}
\alias{winners}
\title{London Marathon winners data set.}
\format{
A tibble with 163 rows and 5 variables:
\describe{
\item{Caegory}{Category of race: Men, Women, Wheelchair Men, or Wheelchair Women}
\item{Year}{Year of race}
\item{Athlete}{Name of winner}
\item{Nationality}{Nationality of winner}
\item{Time}{Winning time}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_winners_of_the_London_Marathon}
}
\usage{
winners
}
\description{
A data set containing information on winners of London Marathon.
}
\keyword{datasets}
